move common IPvXAddressHelper.cs pieces to Common
